Finished styling sign in and sign up components
