+++ src/App.js
@@ -2,7 +2,8 @@ import NavBar from "./Components/NavBar/NavBar";
 import FindLeague from "./routes/find-league/find-league";
 import Home from "./routes/home/home";
 import NextMatches from "./routes/next-matches/nextMatches.component";
-import Authentication from "./routes/authentication/authentication.component";
+import SignInForm from "./routes/sign-in-form/sign-in-form.component";
+import SignUpForm from "./routes/sign-up-form/sign-up-form.component";
 
 import { Route, Routes } from "react-router-dom";
 
@@ -16,7 +17,8 @@ const App = () => {
           <Route index element={<Home />} />
           <Route path="find-league" element={<FindLeague />} />
           <Route path="next-matches" element={<NextMatches />} />
-          <Route path="auth" element={<Authentication />} />
+          <Route path="sign-in" element={<SignInForm />} />
+          <Route path="sign-up" element={<SignUpForm />} />
         </Route>
       </Routes>
     </div>
