--- src/Components/sign-in-form/sign-in-form.styles.scss
@@ -1,35 +1,0 @@
-.login-form {
-  min-width: 250px;
-
-  .sign-in-title {
-    display: flex;
-    justify-content: center;
-    margin-bottom: 20px;
-    padding: 5px;
-
-    h3 {
-      margin: 0;
-      font-family: Arial, Helvetica, sans-serif;
-    }
-  }
-
-  .sign-in-buttons {
-    display: flex;
-    justify-content: center;
-    width: 100%;
-
-    .sign-in-button {
-      font-size: 13px;
-      font-weight: 600;
-      padding: 0 10px;
-      margin: 10px;
-    }
-
-    .sign-with-google-button {
-      font-size: 13px;
-      padding: 0 10px;
-      margin: 10px;
-      font-weight: 600;
-    }
-  }
-}
