+++ src/routes/sign-up-form/sign-up-form.styles.scss
@@ -1,8 +1,11 @@
-.authentication-container {
+.sign-up-container {
   display: flex;
   justify-content: center;
   align-items: center;
-  position: relative;
   height: 100vh;
   width: 100vw;
+
+  h3 {
+    text-align: center;
+  }
 }
